evm-agents/requirements.txt
# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
gunicorn>=21.2.0
pydantic>=2.0.0

# Authentication & Environment
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1
openai>=1.0.0

# HTTP & Networking
requests>=2.31.0
httpx>=0.24.1

# Monitoring & Performance
prometheus-fastapi-instrumentator>=6.0.0
uvicorn-prometheus-logger>=0.2.0

# Validation & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
tenacity>=8.2.3
pyyaml>=6.0.1
